CustomBackGroundViewクラスの作成
